Extract all helper methods into smoke-test-utils and add version to smoke test application
