+++ playwright.config.ts
@@ -5,7 +5,7 @@ const { devices } = require('@playwright/test');
 const config = {
     forbidOnly: !!process.env.CI,
     reporter: 'list',
-    workers: process.env.CI ? 6 : undefined,
+    workers: process.env.CI ? 4 : undefined,
     testDir: 'src/__smoke-test__',
     retries: process.env.CI ? 2 : 2,
     timeout: 300000,
