+++ src/__smoke-test__/ingestion-integ.spec.ts
@@ -1,9 +1,5 @@
 import { test, expect } from '@playwright/test';
-import {
-    RUMClient,
-    GetAppMonitorDataCommand,
-    GetAppMonitorDataCommandInput
-} from '@aws-sdk/client-rum';
+import { RUMClient } from '@aws-sdk/client-rum';
 import {
     HTTP_EVENT_TYPE,
     JS_ERROR_EVENT_TYPE,
@@ -14,86 +10,24 @@ import {
     PAGE_VIEW_EVENT_TYPE,
     SESSION_START_EVENT_TYPE
 } from '../plugins/utils/constant';
+import {
+    getEventIds,
+    getEventsByType,
+    getUrl,
+    isDataPlaneRequest,
+    verifyIngestionWithRetry
+} from 'test-utils/smoke-test-utils';
 
 // Environment variables set through CLI command
 const ENDPOINT = process.env.ENDPOINT;
 const MONITOR_ID = process.env.MONITOR;
-const TEST_URL = process.env.URL || 'http://localhost:9000/smoke_local.html';
+const TEST_URL = getUrl(process.env.URL, process.env.VERSION);
 const MONITOR_NAME = process.env.NAME;
-
 const REGION = ENDPOINT.split('.')[2];
 const TARGET_URL = ENDPOINT + MONITOR_ID + '/';
 
 // Parse region from endpoint
 const rumClient = new RUMClient({ region: REGION });
-
-function getEventsByType(requestBody, eventType) {
-    return requestBody.RumEvents.filter((e) => e.type === eventType);
-}
-function getEventIds(events) {
-    return events.map((e) => e.id);
-}
-
-/**
- * Returns true if the request is a successful PutRumEvents request
- */
-function isDataPlaneRequest(response): boolean {
-    const request = response.request();
-    return (
-        request.method() === 'POST' &&
-        response.status() === 200 &&
-        response.url() === TARGET_URL
-    );
-}
-
-/** Returns true when all events were ingested */
-async function verifyIngestionWithRetry(eventIds, timestamp, retryCount) {
-    while (true) {
-        if (retryCount === 0) {
-            console.log('Retry attempt exhausted.');
-            return false;
-        }
-        try {
-            await isEachEventIngested(eventIds, timestamp);
-            return true;
-        } catch (error) {
-            retryCount -= 1;
-            console.log(`${error.message} Waiting for next retry.`);
-            await new Promise((r) => setTimeout(r, 60000));
-        }
-    }
-}
-
-async function isEachEventIngested(eventIds, timestamp) {
-    let ingestedEvents = new Set();
-    const input: GetAppMonitorDataCommandInput = {
-        Name: MONITOR_NAME,
-        TimeRange: {
-            After: timestamp
-        }
-    };
-    let command = new GetAppMonitorDataCommand(input);
-    // Running tests in parallel require pagination logic, as several test cases have the same timestamp
-    while (true) {
-        const data = await rumClient.send(command);
-        for (let i = 0; i < data.Events.length; i++) {
-            ingestedEvents.add(JSON.parse(data.Events[i]).event_id);
-        }
-        if (data.NextToken) {
-            input.NextToken = data.NextToken;
-            command = new GetAppMonitorDataCommand(input);
-        } else {
-            // If there are no more pages, we can finish the loop
-            break;
-        }
-    }
-
-    for (let i = 0; i < eventIds.length; i++) {
-        if (!ingestedEvents.has(eventIds[i])) {
-            throw new Error(`Event ${eventIds[i]} not ingested.`);
-        }
-    }
-}
 
 // Run the tests in parallel
 test.describe.configure({ mode: 'parallel' });
@@ -107,7 +41,7 @@ test('when session start event is sent then event is ingested', async ({
 
     // Test will timeout if no successful dataplane request is found
     const response = await page.waitForResponse(async (response) =>
-        isDataPlaneRequest(response)
+        isDataPlaneRequest(response, TARGET_URL)
     );
 
     // Parse payload to verify event count
@@ -117,8 +51,10 @@ test('when session start event is sent then event is ingested', async ({
     const eventIds = getEventIds(session);
 
     const isIngestionCompleted = await verifyIngestionWithRetry(
+        rumClient,
         eventIds,
         timestamp,
+        MONITOR_NAME,
         5
     );
     expect(isIngestionCompleted).toEqual(true);
@@ -132,7 +68,7 @@ test('when resource event is sent then event is ingested', async ({ page }) => {
 
     // Test will timeout if no successful dataplane request is found
     const response = await page.waitForResponse(async (response) =>
-        isDataPlaneRequest(response)
+        isDataPlaneRequest(response, TARGET_URL)
     );
 
     // Parse payload to verify event count
@@ -145,8 +81,10 @@ test('when resource event is sent then event is ingested', async ({ page }) => {
     const eventIds = getEventIds(resource);
 
     const isIngestionCompleted = await verifyIngestionWithRetry(
+        rumClient,
         eventIds,
         timestamp,
+        MONITOR_NAME,
         5
     );
     expect(isIngestionCompleted).toEqual(true);
@@ -162,7 +100,7 @@ test('when LCP event is sent then event is ingested', async ({ page }) => {
 
     // Test will timeout if no successful dataplane request is found
     const response = await page.waitForResponse(async (response) =>
-        isDataPlaneRequest(response)
+        isDataPlaneRequest(response, TARGET_URL)
     );
 
     // Parse payload to verify event count
@@ -173,8 +111,10 @@ test('when LCP event is sent then event is ingested', async ({ page }) => {
 
     expect(eventIds.length).not.toEqual(0);
     const isIngestionCompleted = await verifyIngestionWithRetry(
+        rumClient,
         eventIds,
         timestamp,
+        MONITOR_NAME,
         5
     );
     expect(isIngestionCompleted).toEqual(true);
@@ -190,7 +130,7 @@ test('when FID event is sent then event is ingested', async ({ page }) => {
 
     // Test will timeout if no successful dataplane request is found
     const response = await page.waitForResponse(async (response) =>
-        isDataPlaneRequest(response)
+        isDataPlaneRequest(response, TARGET_URL)
     );
 
     // Parse payload to verify event count
@@ -201,8 +141,10 @@ test('when FID event is sent then event is ingested', async ({ page }) => {
 
     expect(eventIds.length).not.toEqual(0);
     const isIngestionCompleted = await verifyIngestionWithRetry(
+        rumClient,
         eventIds,
         timestamp,
+        MONITOR_NAME,
         5
     );
     expect(isIngestionCompleted).toEqual(true);
@@ -220,7 +162,7 @@ test('when navigation events are sent then events are ingested', async ({
 
     // Test will timeout if no successful dataplane request is found
     const response = await page.waitForResponse(async (response) =>
-        isDataPlaneRequest(response)
+        isDataPlaneRequest(response, TARGET_URL)
     );
 
     // Parse payload to verify event count
@@ -235,8 +177,10 @@ test('when navigation events are sent then events are ingested', async ({
     // One initial load, one route change
     expect(eventIds.length).toEqual(2);
     const isIngestionCompleted = await verifyIngestionWithRetry(
+        rumClient,
         eventIds,
         timestamp,
+        MONITOR_NAME,
         5
     );
     expect(isIngestionCompleted).toEqual(true);
@@ -254,7 +198,7 @@ test('when page view event is sent then the event is ingested', async ({
 
     // Test will timeout if no successful dataplane request is found
     const response = await page.waitForResponse(async (response) =>
-        isDataPlaneRequest(response)
+        isDataPlaneRequest(response, TARGET_URL)
     );
 
     // Parse payload to verify event count
@@ -266,8 +210,10 @@ test('when page view event is sent then the event is ingested', async ({
     // One initial load, one route change
     expect(eventIds.length).toEqual(2);
     const isIngestionCompleted = await verifyIngestionWithRetry(
+        rumClient,
         eventIds,
         timestamp,
+        MONITOR_NAME,
         5
     );
     expect(isIngestionCompleted).toEqual(true);
@@ -289,7 +235,7 @@ test('when error events are sent then the events are ingested', async ({
 
     // Test will timeout if no successful dataplane request is found
     const response = await page.waitForResponse(async (response) =>
-        isDataPlaneRequest(response)
+        isDataPlaneRequest(response, TARGET_URL)
     );
 
     // Parse payload to verify event count
@@ -301,8 +247,10 @@ test('when error events are sent then the events are ingested', async ({
     // Expect three js error events
     expect(eventIds.length).toEqual(3);
     const isIngestionCompleted = await verifyIngestionWithRetry(
+        rumClient,
         eventIds,
         timestamp,
+        MONITOR_NAME,
         5
     );
     expect(isIngestionCompleted).toEqual(true);
@@ -322,7 +270,7 @@ test('when http events are sent then the events are ingested', async ({
 
     // Test will timeout if no successful dataplane request is found
     const response = await page.waitForResponse(async (response) =>
-        isDataPlaneRequest(response)
+        isDataPlaneRequest(response, TARGET_URL)
     );
 
     // Parse payload to verify event count
@@ -334,8 +282,10 @@ test('when http events are sent then the events are ingested', async ({
     // Expect three js error events
     expect(eventIds.length).toEqual(2);
     const isIngestionCompleted = await verifyIngestionWithRetry(
+        rumClient,
         eventIds,
         timestamp,
+        MONITOR_NAME,
         5
     );
     expect(isIngestionCompleted).toEqual(true);
